Move IconChip component to ComplianceTable dir
