--- packages/airview-compliance-ui/src/features/icon-chip/index.js
@@ -1,1 +1,0 @@
-export { IconChip } from "./icon-chip";
